coding insert , update screen C

+++ nts.uk/ctx/pr/report/nts.uk.ctx.pr.report.app/src/main/java/nts/uk/ctx/pr/report/app/payment/comparing/command/BaseEntityCommand.java
@@ -1,10 +1,12 @@
 package nts.uk.ctx.pr.report.app.payment.comparing.command;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class BaseEntityCommand {
 	 private String itemCode;
 	 private int categoryAtr;
-	 private int dispOrder;
+	 private BigDecimal dispOrder;
 }
